feat: add shared next.js config
