+++ apps/shop-abc/next.config.mjs
@@ -1,7 +1,7 @@
-// apps/shop-adc/next.config.mjs
+// apps/shop-abc/next.config.mjs
 import path from "node:path";
 import { fileURLToPath } from "node:url";
-import base from "../cms/next.config.mjs";
+import { baseConfig, withShopCode } from "@acme/next-config";
 
 /* ------------------------------------------------------------------ */
 /*  ES-module-safe __dirname                                           */
@@ -9,16 +9,13 @@ import base from "../cms/next.config.mjs";
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
 
-/** @type {import('next').NextConfig} */
-export default {
-  ...base,
-
+export default withShopCode(process.env.SHOP_CODE, {
   /* 1️⃣ ‒ don’t fail the build on lint errors */
   eslint: { ignoreDuringBuilds: true },
 
   webpack(config, { isServer }) {
-    if (typeof base.webpack === "function") {
-      config = base.webpack(config, { isServer });
+    if (typeof baseConfig.webpack === "function") {
+      config = baseConfig.webpack(config, { isServer });
     }
 
     /* 3️⃣ – same alias set as Template App ------------------------------- */
@@ -32,9 +29,4 @@ export default {
 
     return config;
   },
-
-  env: {
-    ...base.env,
-    SHOP_CODE: "abc",
-  },
-};
+});
